pyproject.toml
[project]
name = "yt-api"
version = "0.1.0"
description = ""
authors = [
    {name = "henwix",email = "492b672abx1@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (==5.1.6)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.4.0)",
    "django-storages (==1.14.5)",
    "boto3 (==1.37.6)",
    "celery (==5.4.0)",
    "django-celery-beat (==2.7.0)",
    "django-celery-results (==2.5.1)",
    "django-redis (==5.4.0)",
    "djoser (==2.3.1)",
    "drf-spectacular (==0.28.0)",
    "django-filter (==25.1)",
    "punq (==0.7.0)",
    "pillow (==11.1.0)",
    "django-extensions (==3.2.3)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-prometheus (>=2.3.1,<3.0.0)",
    "psycopg2 (==2.9.10)",
    "orjson (>=3.10.16,<4.0.0)",
    "flower (>=2.0.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
ipython = "8.32.0"
pytest-django = "4.10.0"
factory-boy = "3.3.3"
isort = "^6.0.1"
pre-commit = "^4.2.0"
django-silk = "5.3.2"
tqdm = "^4.67.1"



[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings.prod"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
    ".pytest_cache",
]
known_django=["django","rest_framework"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]